--- Electron-Frontend/app/src/index.html
@@ -1,18 +1,0 @@
-<!DOCTYPE html>
-<html lang="de">
-<head>
-    <meta charset="UTF-8">
-    <title>Hello World</title>
-    <meta http-equiv="Content-Security-Policy" content="script-src 'self' 'unsafe-inline';"/>
-</head>
-<body>
-<h1>Hello World!</h1>
-We are using node
-<script>document.write(process.versions.node)</script>
-, Chrome
-<script>document.write(process.versions.chrome)</script>
-, and Electron
-<script>document.write(process.versions.electron)</script>
-.
-</body>
-</html>
