--- Electron-Frontend/app/src/main.js
@@ -1,28 +1,0 @@
-const {app, BrowserWindow} = require('electron')
-
-function createWindow() {
-    const win = new BrowserWindow({
-        width: 800,
-        height: 600,
-        webPreferences: {
-            nodeIntegration: true
-        }
-    })
-
-    win.loadFile('app/src/index.html')
-    win.webContents.openDevTools()
-}
-
-app.whenReady().then(createWindow)
-
-app.on('window-all-closed', () => {
-    if (process.platform !== 'darwin') {
-        app.quit()
-    }
-})
-
-app.on('activate', () => {
-    if (BrowserWindow.getAllWindows().length === 0) {
-        createWindow()
-    }
-})
