+++ Electron-Frontend/src/login.html
@@ -1,0 +1,26 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <title>Login</title>
+    <link rel="stylesheet" href="node_modules/bootstrap/dist/css/bootstrap.min.css" />
+    <script>let $ = require('jquery');</script>
+    <script>require('popper.js');</script>
+    <script>require('bootstrap');</script>
+</head>
+<body>
+    <div class="container">
+        <br/>
+        <div class="form-group">
+            <label for="txtUsr">User Name:</label>
+            <input type="text" class="form-control" id="txtUsr">
+        </div>
+        <div class="form-group">
+            <label for="txtPwd">Password:</label>
+            <input type="password" class="form-control" id="txtPwd">
+        </div>
+        <button class="btn btn-default" id="btn-login">Login</button>
+        <label  id="lbl"></label>
+    </div>
+</body>
+</html>
